# Commit message
Merge completion results into NodeProcessor

In this commit, I made the following changes:

- Removed the import statement for "github.com/hashicorp/go-multierror"
- Refactored the Step method in the NodeProcessor struct:
  - Removed the TODO comment and split the function into smaller functions
  - Extracted the code block parsing logic into a separate function called parseCodeBlock
  - Removed the orphanSnippet0 function
  - Replaced the use of multierror with a single err variable
  - Refactored the merging of completion results into a new method called mergeCompletionResults
  - Updated the code to use the new mergeCompletionResults method
- Added documentation for the mergeCompletionResults method
- Fixed various code formatting and minor issues

The objective behind these changes was to improve the organization and readability of the NodeProcessor code. By splitting the Step method into smaller functions and removing the orphanSnippet0 function, the code became more modular and easier to understand. The mergeCompletionResults method was introduced to encapsulate the merging logic and make it reusable.