# Commit Message
[Refactor] Add Query function to OnlineIndex struct

- Added a new function `Query` to the `OnlineIndex` struct in the `index.go` file.
- The `Query` function takes a query embedding and performs a search in the online index.
- The function returns a list of hits, each containing the corresponding entry from the index and the distance between the query and the entry embedding.
- This change aims to provide a convenient way to query the online index and get relevant results.

# Reasoning
The new `Query` function will enhance the functionality of the `OnlineIndex` struct by allowing users to perform searches in the online index. With this addition, users can pass a query embedding and obtain a list of hits that match the query. This will provide better search capabilities and improve the overall usefulness of the `OnlineIndex` struct.