# Commit Message
Add Node interface and update NodeBase struct

- Added a new interface `Node` which defines the methods for a node in the graph.
- Updated the `NodeBase` struct to implement the `Node` interface.
- Added methods to retrieve the ID and UUID of the node.
- Added methods to get the parent node, set the parent node, get the children nodes, get the edges, check if the node is a container or leaf node, and get the comments.
- Implemented methods to attach and detach the node to/from a graph.
- Implemented methods to add, remove, and replace child nodes.
- Implemented methods to set and unset edges, and get the edge for a given key.
- Removed the unused `detachFromGraph` method.
- Renamed the `DetachFromGraph` method to `detachFromGraph`.

The objective of these changes is to provide a standardized interface (`Node`) for working with graph nodes and to enhance the functionality of the `NodeBase` struct by implementing the interface methods. These changes enable better abstraction and encapsulation of the graph node logic, making the code more modular, extensible, and maintainable.